+++ src/event-dashboards/index.js
@@ -9,14 +9,14 @@ import WorkerStatusPage from './WorkerStatusPage';
 import RawLogsPage from './RawLogsPage';
 
 // Mount Workers dashboard.
-const workersMount = document.getElementById( 'event-logger-workers' );
+const workersMount = document.getElementById( 'newspack-nodes-workers' );
 if ( workersMount ) {
 	const root = createRoot( workersMount );
 	root.render( <WorkerStatusPage /> );
 }
 
 // Mount Raw Logs dashboard.
-const rawlogsMount = document.getElementById( 'event-logger-rawlogs' );
+const rawlogsMount = document.getElementById( 'newspack-nodes-rawlogs' );
 if ( rawlogsMount ) {
 	const root = createRoot( rawlogsMount );
 	root.render( <RawLogsPage /> );